# Add LRU cache to RequirementsFinder._get_names