@Tenzer 2019-02-28
Looking at the errors from the AppVeyor build, it doesn't seem like many of the are related to this PR. Are those tests expected to pass?