@timothycrosley 2019-03-02
This looks great, Thanks!

AppVeyor stopped running at some point - unintentionally, and as a result, a lot of Windows bugs were introduced. I spent some time over the last few days fixing those after turning it back on.